adjust the symbol exposure
